+++ ionoscloud/import_k8s_node_pool_test.go
@@ -4,7 +4,6 @@ package ionoscloud
 
 import (
 	"fmt"
-	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -13,16 +12,6 @@ import (
 
 func TestAcck8sNodepool_ImportBasic(t *testing.T) {
 	resourceName := "terraform_acctest"
-	publicIp1 := os.Getenv("TF_ACC_IONOS_PUBLIC_IP_1")
-	if publicIp1 == "" {
-		t.Errorf("TF_ACC_IONOS_PUBLIC_1 not set; please set it to a valid public IP for the us/las zone")
-		t.FailNow()
-	}
-	publicIp2 := os.Getenv("TF_ACC_IONOS_PUBLIC_IP_2")
-	if publicIp2 == "" {
-		t.Errorf("TF_ACC_IONOS_PUBLIC_2 not set; please set it to a valid public IP for the us/las zone")
-		t.FailNow()
-	}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t) },
@@ -30,7 +19,7 @@ func TestAcck8sNodepool_ImportBasic(t *testing.T) {
 		CheckDestroy:      testAccCheckk8sNodepoolDestroyCheck,
 		Steps: []resource.TestStep{
 			{
-				Config: fmt.Sprintf(testAccCheckk8sNodepoolConfigBasic, resourceName, publicIp1, publicIp2),
+				Config: fmt.Sprintf(testAccCheckk8sNodepoolConfigBasic, resourceName),
 			},
 			{
 				ResourceName:            fmt.Sprintf("ionoscloud_k8s_node_pool.%s", resourceName),